+++ src/fides/api/models/privacy_preference.py
@@ -162,22 +162,30 @@ class ConsentReportingMixin:
     feature = Column(
         Integer, index=True
     )  # When saving privacy preferences with respect to a TCF feature directly
-    purpose = Column(
+    purpose_consent = Column(
         Integer, index=True
-    )  # When saving privacy preferences with respect to a TCF purpose directly
+    )  # When saving privacy preferences with respect to a TCF purpose with a consent legal basis
+    purpose_legitimate_interests = Column(
+        Integer, index=True
+    )  # When saving privacy preferences with respect to a TCF purpose with a legitimate interests legal basis
     special_feature = Column(
         Integer, index=True
     )  # When saving privacy preferences with respect to a TCF special feature directly
     special_purpose = Column(
         Integer, index=True
     )  # When saving privacy preferences with respect to a TCF special purpose directly
-    vendor = Column(
+    vendor_consent = Column(
         String, index=True
-    )  # When saving privacy preferences with respect to a vendor directly. Vendors can apply to multiple systems.
-    system = Column(
+    )  # When saving privacy preferences with respect to a vendor with a legal basis of consent
+    vendor_legitimate_interests = Column(
         String, index=True
-    )  # When saving privacy preferences with respect to a system id directly, in the case where the vendor is unknown
-
+    )  # When saving privacy preferences with respect to a vendor with a legal basis of legitimate interests
+    system_consent = Column(
+        String, index=True
+    )  # When saving privacy preferences with respect to a system id with consent legal basis, in the case where the vendor is unknown
+    system_legitimate_interests = Column(
+        String, index=True
+    )  # When saving privacy preferences with respect to a system id with legitimate interests legal basis, in the case where the vendor is unknown
     tcf_version = Column(String)
 
     @property
@@ -207,10 +215,13 @@ class ConsentReportingMixin:
         or served against based on which field exists"""
         consent_record_type_mapping = {
             "privacy_notice_history_id": ConsentRecordType.privacy_notice_id,
-            "purpose": ConsentRecordType.purpose,
+            "purpose_consent": ConsentRecordType.purpose_consent,
+            "purpose_legitimate_interests": ConsentRecordType.purpose_legitimate_interests,
             "special_purpose": ConsentRecordType.special_purpose,
-            "vendor": ConsentRecordType.vendor,
-            "system": ConsentRecordType.system,
+            "vendor_consent": ConsentRecordType.vendor_consent,
+            "vendor_legitimate_interests": ConsentRecordType.vendor_legitimate_interests,
+            "system_consent": ConsentRecordType.system_consent,
+            "system_legitimate_interests": ConsentRecordType.system_legitimate_interests,
             "special_feature": ConsentRecordType.special_feature,
             "feature": ConsentRecordType.feature,
         }
@@ -260,11 +271,14 @@ def _validate_before_saving_consent_history(
             "Must supply a verified provided identity id or a fides_user_device_provided_identity_id"
         )
 
-    if data.get("system"):
-        system = db.query(System).filter(System.id == data.get("system")).first()
+    system_id: Optional[str] = data.get("system_consent") or data.get(
+        "system_legitimate_interests"
+    )
+    if system_id:
+        system = db.query(System).filter(System.id == system_id).first()
         if not system:
             raise SystemNotFound(
-                f"Can't save consent against invalid system id '{data.get('system')}'."
+                f"Can't save consent against invalid system id '{system_id}'."
             )
 
     tcf_attributes: Dict[str, Union[Optional[str], Optional[int]]] = {
@@ -594,7 +608,12 @@ class LastSavedMixin:
     # ==== TCF Attributes that can be served ==== #
     feature = Column(Integer, index=True)  # When a feature was served directly (TCF)
 
-    purpose = Column(Integer, index=True)  # When a purpose was served directly (TCF)
+    purpose_consent = Column(
+        Integer, index=True
+    )  # When a TCF purpose with consent legal basis was served directly (TCF)
+    purpose_legitimate_interests = Column(
+        Integer, index=True
+    )  # When a TCF purpose with legitimate interests legal basis was served directly (TCF)
 
     special_feature = Column(
         Integer, index=True
@@ -604,13 +623,19 @@ class LastSavedMixin:
         Integer, index=True
     )  # When a special purpose was served directly (TCF)
 
-    vendor = Column(
+    vendor_consent = Column(
         String, index=True
-    )  # When a vendor was served directly (TCF). Vendors can apply to multiple systems.
+    )  # When a TCF vendor with consent legal basis was served directly (TCF)
+    vendor_legitimate_interests = Column(
+        String, index=True
+    )  # When a TCF vendor with legitimate interest legal basis was served directly (TCF)
 
-    system = Column(
+    system_consent = Column(
         String, index=True
-    )  # The id of a system (TCF). Used for when the specific vendor type is unknown.
+    )  # When a system id with consent legal basis was served directly.  Used for when the specific vendor type is unknown.
+    system_legitimate_interests = Column(
+        String, index=True
+    )  # When a system id with legitimate interest legal basis was served directly.  Used for when the specific vendor type is unknown.
 
     @declared_attr
     def provided_identity_id(cls) -> Column:
@@ -686,11 +711,23 @@ class CurrentPrivacyPreference(LastSavedMixin, Base):
             "privacy_notice_id",
             name="fides_user_device_identity_privacy_notice",
         ),
-        UniqueConstraint("provided_identity_id", "purpose", name="identity_purpose"),
+        UniqueConstraint(
+            "provided_identity_id", "purpose_consent", name="identity_purpose_consent"
+        ),
+        UniqueConstraint(
+            "provided_identity_id",
+            "purpose_legitimate_interests",
+            name="identity_purpose_leg_interests",
+        ),
         UniqueConstraint(
             "fides_user_device_provided_identity_id",
-            "purpose",
-            name="fides_user_device_identity_purpose",
+            "purpose_consent",
+            name="fides_user_device_identity_purpose_consent",
+        ),
+        UniqueConstraint(
+            "fides_user_device_provided_identity_id",
+            "purpose_legitimate_interests",
+            name="fides_user_device_identity_purpose_leg_interests",
         ),
         UniqueConstraint(
             "provided_identity_id", "special_purpose", name="identity_special_purpose"
@@ -700,17 +737,41 @@ class CurrentPrivacyPreference(LastSavedMixin, Base):
             "special_purpose",
             name="fides_user_device_identity_special_purpose",
         ),
-        UniqueConstraint("provided_identity_id", "vendor", name="identity_vendor"),
         UniqueConstraint(
-            "fides_user_device_provided_identity_id",
-            "vendor",
-            name="fides_user_device_identity_vendor",
+            "provided_identity_id", "vendor_consent", name="identity_vendor_consent"
         ),
-        UniqueConstraint("provided_identity_id", "system", name="identity_system"),
+        UniqueConstraint(
+            "provided_identity_id",
+            "vendor_legitimate_interests",
+            name="identity_vendor_leg_interests",
+        ),
         UniqueConstraint(
             "fides_user_device_provided_identity_id",
-            "system",
-            name="fides_user_device_identity_system",
+            "vendor_consent",
+            name="fides_user_device_identity_vendor_consent",
+        ),
+        UniqueConstraint(
+            "fides_user_device_provided_identity_id",
+            "vendor_legitimate_interests",
+            name="fides_user_device_identity_vendor_leg_interests",
+        ),
+        UniqueConstraint(
+            "provided_identity_id", "system_consent", name="identity_system_consent"
+        ),
+        UniqueConstraint(
+            "provided_identity_id",
+            "system_legitimate_interests",
+            name="identity_system_leg_interests",
+        ),
+        UniqueConstraint(
+            "fides_user_device_provided_identity_id",
+            "system_consent",
+            name="fides_user_device_identity_system_consent",
+        ),
+        UniqueConstraint(
+            "fides_user_device_provided_identity_id",
+            "system_legitimate_interests",
+            name="fides_user_device_identity_system_leg_interests",
         ),
         UniqueConstraint("provided_identity_id", "feature", name="identity_feature"),
         UniqueConstraint(
@@ -782,13 +843,23 @@ class LastServedNotice(LastSavedMixin, Base):
         ),
         UniqueConstraint(
             "provided_identity_id",
-            "purpose",
-            name="last_served_identity_purpose",
+            "purpose_consent",
+            name="last_served_identity_purpose_consent",
+        ),
+        UniqueConstraint(
+            "provided_identity_id",
+            "purpose_legitimate_interests",
+            name="last_served_identity_purpose_legitimate_interests",
         ),
         UniqueConstraint(
             "fides_user_device_provided_identity_id",
-            "purpose",
-            name="last_served_fides_user_device_identity_purpose",
+            "purpose_consent",
+            name="last_served_fides_user_device_identity_purpose_consent",
+        ),
+        UniqueConstraint(
+            "fides_user_device_provided_identity_id",
+            "purpose_legitimate_interests",
+            name="last_served_fides_user_device_identity_purpose_leg_interests",
         ),
         UniqueConstraint(
             "provided_identity_id",
@@ -812,23 +883,43 @@ class LastServedNotice(LastSavedMixin, Base):
         ),
         UniqueConstraint(
             "provided_identity_id",
-            "vendor",
-            name="last_served_identity_vendor",
-        ),
-        UniqueConstraint(
-            "fides_user_device_provided_identity_id",
-            "vendor",
-            name="last_served_fides_user_device_identity_vendor",
+            "vendor_consent",
+            name="last_served_identity_vendor_consent",
         ),
         UniqueConstraint(
             "provided_identity_id",
-            "system",
-            name="last_served_identity_system",
+            "vendor_legitimate_interests",
+            name="last_served_identity_vendor_leg_interests",
         ),
         UniqueConstraint(
             "fides_user_device_provided_identity_id",
-            "system",
-            name="last_served_fides_user_device_identity_system",
+            "vendor_consent",
+            name="last_served_fides_user_device_identity_vendor_consent",
+        ),
+        UniqueConstraint(
+            "fides_user_device_provided_identity_id",
+            "vendor_legitimate_interests",
+            name="last_served_fides_user_device_identity_vendor_leg_interests",
+        ),
+        UniqueConstraint(
+            "provided_identity_id",
+            "system_consent",
+            name="last_served_identity_system_consent",
+        ),
+        UniqueConstraint(
+            "provided_identity_id",
+            "system_legitimate_interests",
+            name="last_served_identity_system_leg_interests",
+        ),
+        UniqueConstraint(
+            "fides_user_device_provided_identity_id",
+            "system_consent",
+            name="last_served_fides_user_device_identity_system_consent",
+        ),
+        UniqueConstraint(
+            "fides_user_device_provided_identity_id",
+            "system_legitimate_interests",
+            name="last_served_fides_user_device_identity_system_leg_interests",
         ),
         UniqueConstraint(
             "provided_identity_id",